+++ src/enhancedScatterChartSettingsModel.ts
@@ -1,11 +1,13 @@
-import powerbiVisualsApi from "powerbi-visuals-api";
+import powerbi from "powerbi-visuals-api";
 import { formattingSettings } from "powerbi-visuals-utils-formattingmodel";
 import { legendInterfaces } from "powerbi-visuals-utils-chartutils";
 import { ColorHelper } from "powerbi-visuals-utils-colorutils";
 import { LegendDataPoint } from "powerbi-visuals-utils-chartutils/lib/legend/legendInterfaces";
 
-import ISelectionId = powerbiVisualsApi.visuals.ISelectionId;
+import ISelectionId = powerbi.visuals.ISelectionId;
 import LegendPosition = legendInterfaces.LegendPosition;
+
+import ILocalizationManager = powerbi.extensibility.ILocalizationManager;
 
 import Card = formattingSettings.Card;
 import Model = formattingSettings.Model;
@@ -13,33 +15,15 @@ import Model = formattingSettings.Model;
 import IEnumMember = powerbi.IEnumMember;
 import { EnhancedScatterChartDataPoint } from "./dataInterfaces";
 
-export enum DisplayUnitsType {
-    Auto = 0,
-    None = 1,
-    Thousands = 1000,
-    Millions = 1000000,
-    Billions = 1000000000,
-    Triilions = 1000000000000
-}
-
-const displayUnitsOptions : IEnumMember[] = [
-    {value: DisplayUnitsType.Auto, displayName : DisplayUnitsType[DisplayUnitsType.Auto]},
-    {value: DisplayUnitsType.None, displayName : DisplayUnitsType[DisplayUnitsType.None]},
-    {value: DisplayUnitsType.Thousands, displayName : DisplayUnitsType[DisplayUnitsType.Thousands]},
-    {value: DisplayUnitsType.Millions, displayName : DisplayUnitsType[DisplayUnitsType.Millions]},
-    {value: DisplayUnitsType.Billions, displayName : DisplayUnitsType[DisplayUnitsType.Billions]},
-    {value: DisplayUnitsType.Triilions, displayName : DisplayUnitsType[DisplayUnitsType.Triilions]}
-]
-
 const positionOptions : IEnumMember[] = [
-    {value : LegendPosition[LegendPosition.Top], displayName : LegendPosition[LegendPosition.Top]},
-    {value : LegendPosition[LegendPosition.Bottom], displayName : LegendPosition[LegendPosition.Bottom]},
-    {value : LegendPosition[LegendPosition.Left], displayName : LegendPosition[LegendPosition.Left]},
-    {value : LegendPosition[LegendPosition.Right], displayName : LegendPosition[LegendPosition.Right]},
-    {value : LegendPosition[LegendPosition.TopCenter], displayName : LegendPosition[LegendPosition.TopCenter]},
-    {value : LegendPosition[LegendPosition.BottomCenter], displayName : LegendPosition[LegendPosition.BottomCenter]},
-    {value : LegendPosition[LegendPosition.LeftCenter], displayName : LegendPosition[LegendPosition.LeftCenter]},
-    {value : LegendPosition[LegendPosition.RightCenter], displayName : LegendPosition[LegendPosition.RightCenter]},
+    {value : LegendPosition[LegendPosition.Top], displayName : "Visual_LegendPosition_Top"},
+    {value : LegendPosition[LegendPosition.Bottom], displayName : "Visual_LegendPosition_Bottom"},
+    {value : LegendPosition[LegendPosition.Left], displayName : "Visual_LegendPosition_Left"},
+    {value : LegendPosition[LegendPosition.Right], displayName : "Visual_LegendPosition_Right"},
+    {value : LegendPosition[LegendPosition.TopCenter], displayName : "Visual_LegendPosition_TopCenter"},
+    {value : LegendPosition[LegendPosition.BottomCenter], displayName : "Visual_LegendPosition_BottomCenter"},
+    {value : LegendPosition[LegendPosition.LeftCenter], displayName : "Visual_LegendPosition_LeftCenter"},
+    {value : LegendPosition[LegendPosition.RightCenter], displayName : "Visual_LegendPosition_RightCenter"},
 ];
 
 class FontSizeSettings {
@@ -48,9 +32,13 @@ class FontSizeSettings {
     public static readonly MaxFontSize: number = 60;
 }
 
-export class EnableDataPointCardSettings extends Card {
+class StrokeWidthSettings {
+    public static readonly DefaultStrokeWidth: number = 1;
+    public static readonly MinStrokeWidth: number = 1;
+    public static readonly MaxStrokeWidth: number = 5;
+}
 
-    public strokeWidth: number = 1;
+export class EnableDataPointCardSettings extends Card {
 
     defaultColor = new formattingSettings.ColorPicker({
         name: "defaultColor",
@@ -108,11 +96,10 @@ export class ScatterChartAxisCardSettings extends Card {
         value: undefined
     });
 
-    labelDisplayUnits = new formattingSettings.ItemDropdown({
+    labelDisplayUnits = new formattingSettings.AutoDropdown({
         name: "labelDisplayUnits",
         displayNameKey: "Visual_DisplayUnits",
-        items: displayUnitsOptions,
-        value: displayUnitsOptions[0]
+        value: 0
     });
 
     axisColor = new formattingSettings.ColorPicker({
@@ -191,11 +178,11 @@ export class EnableLegendCardSettings extends Card {
         value: FontSizeSettings.DefaultFontSize,
         options: {
             minValue: {
-                type: powerbiVisualsApi.visuals.ValidatorType.Min,
+                type: powerbi.visuals.ValidatorType.Min,
                 value: FontSizeSettings.MinFontSize,
             },
             maxValue: {
-                type: powerbiVisualsApi.visuals.ValidatorType.Max,
+                type: powerbi.visuals.ValidatorType.Max,
                 value: FontSizeSettings.MaxFontSize,
             }
         }
@@ -227,11 +214,11 @@ export class EnableCategoryLabelsCardSettings extends Card {
         value: FontSizeSettings.DefaultFontSize,
         options: {
             minValue: {
-                type: powerbiVisualsApi.visuals.ValidatorType.Min,
+                type: powerbi.visuals.ValidatorType.Min,
                 value: FontSizeSettings.MinFontSize,
             },
             maxValue: {
-                type: powerbiVisualsApi.visuals.ValidatorType.Max,
+                type: powerbi.visuals.ValidatorType.Max,
                 value: FontSizeSettings.MaxFontSize,
             }
         }
@@ -304,9 +291,25 @@ export class EnableOutlineCardSettings extends Card {
         topLevelToggle: true
     });
 
+    strokeWidth = new formattingSettings.NumUpDown({
+        name: "strokeWidth",
+        displayNameKey: "Visual_StrokeWidth",
+        value: StrokeWidthSettings.DefaultStrokeWidth,
+        options: {
+            minValue: {
+                type: powerbi.visuals.ValidatorType.Min,
+                value: StrokeWidthSettings.MinStrokeWidth,
+            },
+            maxValue: {
+                type: powerbi.visuals.ValidatorType.Max,
+                value: StrokeWidthSettings.MaxStrokeWidth,
+            }
+        }
+    });
+
     name: string = "outline";
     displayNameKey: string = "Visual_Outline";
-    slices = [this.show];
+    slices = [this.show, this.strokeWidth];
 }
 
 export class EnhancedScatterChartSettingsModel extends Model {
@@ -323,6 +326,16 @@ export class EnhancedScatterChartSettingsModel extends Model {
     cards = [this.enableDataPointCardSettings, this.enableCategoryAxisCardSettings, this.enableValueAxisCardSettings,
         this.enableLegendCardSettings, this.enableCategoryLabelsCardSettings, this.enableFillPointCardSettings,
         this.enableBackdropCardSettings, this.enableCrosshairCardSettings, this.enableOutlineCardSettings];
+
+    setLocalizedOptions(localizationManager: ILocalizationManager) { 
+        this.setLocalizedDisplayName(positionOptions, localizationManager);
+    }
+
+    public setLocalizedDisplayName(options: IEnumMember[], localizationManager: ILocalizationManager) {
+        options.forEach(option => {
+            option.displayName = localizationManager.getDisplayName(option.displayName.toString())
+        });
+    }
 
     /**
      * populate colorSelector object categories formatting properties